+++ singleshot/dipole_trap_single_shot.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 
 from analysislib.common.analysis_config import kinetix_system, BulkGasAnalysisConfig
-from analysislib.common.bulk_gas_analysis import BulkGasPreprocessor
+from analysislib.common.bulk_gas_preproc import BulkGasPreprocessor
 from analysislib.common.bulk_gas_statistics import BulkGasStatistician
 from analysislib.common.image import ROI
 from analysislib.common.plot_config import PlotConfig
@@ -14,22 +14,22 @@ analysis_config = BulkGasAnalysisConfig(
     bkg_roi=ROI(xmin=1900, xmax=2400, ymin=1900, ymax=2400),
 )
 
-bulk_gas_analyzer = BulkGasPreprocessor(
+bulk_gas_preproc = BulkGasPreprocessor(
     config=analysis_config,
     load_type='lyse',
     h5_path=None,
 )
 
-processed_results_fname = bulk_gas_analyzer.process_shot()
+processed_results_fname = bulk_gas_preproc.process_shot()
 
 fig = plt.figure(layout = "constrained", figsize = [10,4])
 subfigs = fig.subfigures(nrows=1, ncols=2, wspace=0.07)
 
-bulk_gas_analyzer.show_images(fig = subfigs[0], raw_img_scale = 800)
+bulk_gas_preproc.show_images(fig = subfigs[0], raw_img_scale = 800)
 
 plotter = BulkGasStatistician(
     preproc_h5_path=processed_results_fname,
-    shot_h5_path=bulk_gas_analyzer.h5_path, # Used only for MLOOP
+    shot_h5_path=bulk_gas_preproc.h5_path, # Used only for MLOOP
     plot_config=PlotConfig(),
 )
